+++ proofs/evar_refiner.ml
@@ -37,7 +37,7 @@ let define_and_solve_constraints evk c env evd =
   match
     List.fold_left
       (fun p (pbty,env,t1,t2) -> match p with
-        | Success evd -> Evarconv.evar_conv_x TranspState.full env evd pbty t1 t2
+        | Success evd -> Evarconv.evar_conv_x TransparentState.full env evd pbty t1 t2
 	| UnifFailure _ as x -> x) (Success evd)
       pbs
   with